chore: Refactor import paths to consolidate schema references
- Updated import statements across multiple files to reference the new `schemas` directory instead of the previous `schema` structure.
- Introduced a new `schemas/index.js` file to streamline exports from the `schemas` directory, enhancing code organization and maintainability.
