+++ src/types/index.ts
@@ -1,3 +1,4 @@
+export * from "./schemas/index.js";
 export * from "./extended-schema.js";
 export * from "./proxy.js";
 export * from "./services/context.js";
