+++ src/types/schemas/index.ts
@@ -1,0 +1,1 @@
+export * from "./a2a/index.js";
